requirements.txt
asgiref==3.2.7
astroid==2.2.5
colorama==0.4.1
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
lazy-object-proxy==1.5.0
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
typed-ast==1.4.1
whitenoise==5.1.0
wrapt==1.12.1